refactor: profile image
